Split Dataset and Component models into new file
