+++ snakebids/__init__.py
@@ -3,9 +3,6 @@ __submodules__ = ["core"]
 
 # <AUTOGEN_INIT>
 from snakebids.core import (
-    BidsComponent,
-    BidsDataset,
-    BidsDatasetDict,
     bids,
     filter_list,
     generate_inputs,
@@ -16,9 +13,6 @@ from snakebids.core import (
 )
 
 __all__ = [
-    "BidsComponent",
-    "BidsDataset",
-    "BidsDatasetDict",
     "bids",
     "filter_list",
     "generate_inputs",
